requirements-ui.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
websockets>=11.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
alembic>=1.10.0
email-validator>=2.0.0
